Function Reference

TopoPlots.enclosing_geometry — Function

enclosing_geometry(G::Type{<: Geometry}, positions, enlarge=1.0)

Returns the Geometry of Type G, that best fits all positions. The Geometry can be enlarged by 1.x, so e.g. enclosing_geometry(Circle, positions, 0.1) will return a Circle that encloses all positions with a padding of 10%.

TopoPlots.labels2positions — Function

labels2positions(labels)

Currently only supports 10/20 layout, by looking it up in TopoPlots.CHANNEL_TO_POSITION_10_20.